requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
djangorestframework==3.16.0
fonttools==4.58.0
future==1.0.0
generic==1.1.4
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
Naked==0.1.32
numpy==2.2.5
packaging==25.0
pillow==11.2.1
psycopg==3.2.8
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
response==0.5.0
scipy==1.15.3
serializers==0.2.4
six==1.17.0
sqlparse==0.5.3
status==0.2.5
urllib3==2.4.0
